requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
datasets>=2.12.0
pillow>=9.5.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.65.0
sacrebleu>=2.3.0
nltk>=3.8.0
opencv-python>=4.8.0
accelerate>=0.20.0
wandb>=0.15.0
